fixes handlers
