+++ packages/next/package.json
@@ -30,22 +30,24 @@
       "import": "./src/routes/*.ts",
       "require": "./src/routes/*.ts",
       "types": "./src/routes/*.ts"
-    }
+    },
+    "./dist/routes/*": "./dist/routes/*.js"
   },
   "devDependencies": {
-    "@payloadcms/ui": "workspace:*",
     "@payloadcms/eslint-config": "workspace:*",
+    "@payloadcms/ui": "workspace:*",
     "payload": "workspace:*",
     "sass": "^1.69.5"
   },
   "dependencies": {
-    "jsonwebtoken": "9.0.1"
+    "jsonwebtoken": "9.0.1",
+    "path-to-regexp": "^6.2.1"
   },
   "peerDependencies": {
-    "payload": "^2.0.0",
-    "next": "^14.0.0",
+    "http-status": "1.6.2",
     "i18next": "22.5.1",
-    "http-status": "1.6.2"
+    "next": "^14.0.0",
+    "payload": "^2.0.0"
   },
   "publishConfig": {
     "exports": {
@@ -63,6 +65,11 @@
         "import": "./dist/pages/*.js",
         "require": "./dist/pages/*.js",
         "types": "./dist/pages/*.d.ts"
+      },
+      "./app/*": {
+        "import": "./dist/app/*.js",
+        "require": "./dist/app/*.js",
+        "types": "./dist/app/*.d.ts"
       },
       "./routes/*": {
         "import": "./dist/routes/*.js",
